+++ api/views.py
@@ -1,3 +1,32 @@
 from django.shortcuts import render
+from django.http import JsonResponse
+from django_filters.rest_framework import DjangoFilterBackend
+from rest_framework.generics import ListAPIView
+from rest_framework.filters import OrderingFilter
 
-# Create your views here.
+from .models import Videos
+from .serializers import VideosSerializer
+
+class sortedVideos(ListAPIView):
+    queryset = Videos.objects.all()
+    serializer_class = VideosSerializer
+    filter_backends = [DjangoFilterBackend, OrderingFilter]
+    filterset_fields = [
+            'id',
+            'topic',
+            'updated',
+            'thumbnail',
+            'title',
+            'description',
+            'author',
+            'published',
+            'duration',
+            'likes',
+            'views',
+            'author_pic_url',
+    ]
+    ordering_fields = [
+        'likes',
+        'views',
+        'published',
+    ]
