+++ django_files/settings.py
@@ -38,6 +38,9 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'api',
+    'rest_framework',
+    'django_filters',
+    'django_extensions',
 ]
 
 MIDDLEWARE = [
